+++ src/SeerGdbWidget.h
@@ -53,8 +53,8 @@ class SeerGdbWidget : public QWidget, protected Ui::SeerGdbWidgetForm {
         void                                setExecutableConnectHostPort        (const QString& connectHostPort);
         const QString&                      executableConnectHostPort           () const;
 
-        void                                setExecutableRRHostPort             (const QString& rrHostPort);
-        const QString&                      executableRRHostPort                () const;
+        void                                setExecutableRRTraceDirectory       (const QString& rrTraceDirectory);
+        const QString&                      executableRRTraceDirectory          () const;
 
         void                                setExecutableCoreFilename           (const QString& coreFilename);
         const QString&                      executableCoreFilename              () const;
@@ -173,6 +173,16 @@ class SeerGdbWidget : public QWidget, protected Ui::SeerGdbWidgetForm {
 
         void                                setSeerOutputLogEnabled             (bool flag);
         bool                                isSeerOutputLogEnabled              () const;
+
+        // RR settings.
+        void                                setRRProgram                        (const QString& program);
+        QString                             rrProgram                           () const;
+
+        void                                setRRArguments                      (const QString& arguments);
+        QString                             rrArguments                         () const;
+
+        void                                setRRGdbArguments                   (const QString& arguments);
+        QString                             rrGdbArguments                      () const;
 
         // Editor manager.
         SeerEditorManagerWidget*            editorManager                       ();
@@ -333,15 +343,20 @@ class SeerGdbWidget : public QWidget, protected Ui::SeerGdbWidgetForm {
     private:
         bool                                isGdbRuning                         () const;
         bool                                startGdb                            ();
+        bool                                startGdbRR                          ();
         void                                killGdb                             ();
         void                                createConsole                       ();
         void                                deleteConsole                       ();
         void                                connectConsole                      ();
         void                                disconnectConsole                   ();
+        SeerConsoleWidget*                  console                             ();
         void                                sendGdbInterrupt                    (int signal);
 
         QString                             _gdbProgram;
         QString                             _gdbArguments;
+        QString                             _gdbRRProgram;
+        QString                             _gdbRRArguments;
+        QString                             _gdbRRGdbArguments;
         bool                                _gdbASyncMode;
         bool                                _gdbNonStopMode;
         bool                                _gdbHandleTerminatingException;
@@ -367,7 +382,7 @@ class SeerGdbWidget : public QWidget, protected Ui::SeerGdbWidgetForm {
         QString                             _executableBreakpointFunctionName;
         int                                 _executablePid;
         QString                             _executableConnectHostPort;
-        QString                             _executableRRHostPort;
+        QString                             _executableRRTraceDirectory;
         QString                             _executableCoreFilename;
         QString                             _executableLaunchMode;
         QString                             _executableBreakMode;